requirements.txt
requests>=2.28.0
pypdf>=4.0.0
python-docx>=1.1.0
python-jobspy>=1.1.70
